Virtual Camera
Currently not working OnSceneLoad

I am trying to use a CameraController tied to the virtual camera since our LoadLevel triggers an entirely new scene. For some reason when the new scene loads the virtual camera has an X through it and will not follow the player.

Will try again tomorrow.