Main Cameras
I deleted all of the cameras in the following scenes and copied the camera from the Startscene over to them:

-FriendsHouse
-Path02
-CarePackageField
-Path03

This fixed the issue with the main camera interfering with the boundary of the virtual camera because they now have the CinemachineBrain component on them. These cameras can be deleted once the project is done if necessary, but are needed for now otherwise we cannot see the scenes as we edit them.